% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander_anova.R
\name{pander_anova}
\alias{pander_anova}
\title{Create a cleaner ANOVA model output with pander.}
\usage{
pander_anova(anova_model_name)
}
\arguments{
\item{anova_model_name}{ANOVA model. Model whose output will be cleaned.}
}
\value{
neat_output	pander table. Cleaned ANOVA model output.
}
\description{
Create a cleaner ANOVA model output with pander.
}
