% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander_lm.R
\name{pander_lm}
\alias{pander_lm}
\title{Create a cleaner lm model output with pander.}
\usage{
pander_lm(lm_model_name, stats.caption = FALSE)
}
\description{
Parameters
 ----------
 lm_model_name : lm model
    Model whose output will be cleaned.
}
\details{
stats.caption : boolean, optional (default: FALSE)
    Specify whether or not to provide the adjusted
    R-squared and F-statistics in the table caption.

Output
 ------
 neat_output : table
    Cleaned lm model output.
}
