% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander_lme.R
\name{pander_lme}
\alias{pander_lme}
\title{Create a cleaner lme4 model output with pander.}
\usage{
pander_lme(lme_model_name, stats.caption = FALSE)
}
\description{
Parameters
 ----------
 lme_model_name : lme4 model
    Model whose output will be cleaned.
}
\details{
stats.caption : boolean, optional (default: FALSE)
    Specify whether or not to provide the marginal
    (i.e., fixed effects only) and conditional (i.e.,
    fixed and random effects) R-squared values in the
    table's caption.

Output
 ------
 neat_output : pander table
    Cleaned lme4 model output.
export
}
