% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander_lme_to_latex.R
\name{pander_lme_to_latex}
\alias{pander_lme_to_latex}
\title{Export an LMER summary table to a LaTex file.}
\usage{
pander_lme_to_latex(lme_model_name, save_filename)
}
\description{
Note that this will retain the row index column,
 which will be given a name`neat" in the table.
 This will be need to be manually removed in the
 final `.tex` file.
}
\details{
Parameters
 ----------
 lme_model_name : lme4 model
    Model whose output will be cleaned.

save_filename : str
    File name for saved LaTex model.

Output
 ------
 neat_output : pander table
    Cleaned lme4 model output.
}
