% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable_lme.R
\name{xtable_lme}
\alias{xtable_lme}
\title{Create a cleaner lme4 model output for printing with xtable.}
\usage{
xtable_lme(lme_model_name)
}
\description{
Parameters
 ----------
 lme_model_name : lme4 model
    Model whose output will be cleaned.
}
\details{
Output
 ------
 neat_output : table
    Cleaned lme4 model output.
}
